This commit adds support for clang-format to the project.
I have added a `.clang-format` file to the root of the project, which contains the formatting options for clang-format.
These options control the way that the code is formatted when `clang-format` is run.

I have also created a `clang-format.yml` file in the `.github/workflows` directory,
which sets up a Github Actions workflow that runs `clang-format` on any changes made to the codebase.
This workflow will automatically format any changes made to the codebase and commit them back to the repository.

Please note that this formatting will only be applied to changes made after the workflow has been set up.
Existing code will not be automatically formatted, but can be manually formatted using `clang-format`.